Course-8 HomeTask-1 Update

+++ hrms/src/main/java/kodlama/ht6/hrms/core/dataAccess/abstracts/StaffDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import kodlama.ht6.hrms.core.entities.concretes.Staff;
 
-public interface StaffDao extends JpaRepository<Staff, Integer>{
+public interface StaffDao extends JpaRepository<Staff,Long>{
 
 }
