+++ hrms/src/main/java/kodlama/ht6/hrms/core/dataAccess/abstracts/UnconfirmedUserDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import kodlama.ht6.hrms.core.entities.concretes.UnconfirmedUser;
 
-public interface UnconfirmedUserDao extends JpaRepository<UnconfirmedUser, Integer>{
-	public UnconfirmedUser getByUserId(int userId);
+public interface UnconfirmedUserDao extends JpaRepository<UnconfirmedUser, Long>{
+	public UnconfirmedUser getByUserId(long userId);
 
 }
