+++ hrms/src/main/java/kodlama/ht6/hrms/core/entities/concretes/User.java
@@ -14,6 +14,9 @@ import javax.persistence.JoinTable;
 import javax.persistence.ManyToMany;
 import javax.persistence.Table;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
 import kodlama.ht6.hrms.core.entities.abstracts.User_;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -24,24 +27,27 @@ import lombok.NoArgsConstructor;
 @Entity
 @Table(name = "users")
 @Inheritance(strategy = InheritanceType.JOINED)
+//@JsonIgnoreProperties(value= {"passwordHash", "passwordSalt", "claims"})
 public @Data class User implements User_{
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id", updatable = false, nullable = false)
-	private int id;
+	private long id;
 
 	@Column(name = "email")
 	private String email;
 
+//	@JsonIgnore
 	@Column(name = "password_hash")
 	private byte[] passwordHash;
-
+	
+//	@JsonIgnore
 	@Column(name = "password_salt")
 	private byte[] passwordSalt;
-	
+//	@JsonIgnore	
 	@Column(name = "active")
 	private boolean active;
-
+//	@JsonIgnore
 	@ManyToMany
 	@JoinTable(name = "user_claims",
 		joinColumns = {	@JoinColumn (name= "user_id", referencedColumnName = "id")},
