+++ hrms/src/main/java/kodlama/ht6/hrms/dataAccess/abstracts/JobSeekerDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import kodlama.ht6.hrms.entities.concretes.JobSeeker;
 
-public interface JobSeekerDao extends JpaRepository<JobSeeker, Integer>{
+public interface JobSeekerDao extends JpaRepository<JobSeeker, Long>{
 
 }
